+++ plugins/effect/gamma/gamma.c
@@ -4,7 +4,7 @@
  * This file is part of Enfle.
  *
  * Last Modified: Sat Mar  6 12:00:56 2004.
- * $Id: gamma.c,v 1.5 2004/03/06 03:43:36 sian Exp $
+ * $Id: gamma.c,v 1.6 2008/04/19 09:28:05 sian Exp $
  *
  * Enfle is free software; you can redistribute it and/or modify it
  * under the terms of the GNU General Public License version 2 as
@@ -85,7 +85,7 @@ ENFLE_PLUGIN_EXIT(effect_gamma, p)
 static int
 set_gamma(void *a)
 {
-  gamma_idx = (int)a;
+  gamma_idx = (int)(long)a;
 
   return 1;
 }
